test: add fixtures
